// src/modules/pagos/pago.enums.ts
export enum EstadoPago {
  PENDIENTE = 'pendiente',
  APROBADO = 'aprobado',
  RECHAZADO = 'rechazado',
  CANCELADO = 'cancelado',
  PRESENCIAL_PENDIENTE = 'presencial_pendiente',
  PRESENCIAL_COBRADO = 'presencial_cobrado',
}

export enum MedioPago {
  MERCADO_PAGO = 'mercado_pago',
  PRESENCIAL = 'presencial',
}